+++ permuta/permutils/__init__.py
@@ -2,6 +2,7 @@ from .finite import is_finite
 from .groups import dihedral_group
 from .insertion_encodable import InsertionEncodablePerms
 from .polynomial import PolyPerms
+from .statistics import PermutationStatistic
 from .symmetry import (
     all_symmetry_sets,
     antidiagonal_set,
@@ -41,4 +42,5 @@ __all__ = [
     "rotate_90_clockwise_set",
     "rotate_180_clockwise_set",
     "rotate_270_clockwise_set",
+    "PermutationStatistic",
 ]